extract background rounds to past_rounds module
